+++ src/components/Settings/DarklightModeAnimation.tsx
@@ -1,8 +1,10 @@
-import React, { useEffect, useRef } from 'react';
+import * as Haptics from 'expo-haptics';
+import { includes } from 'lodash';
+import React, { useCallback, useEffect, useRef } from 'react';
 import { ViewStyle } from 'react-native';
 import Rive, { Alignment, Fit, type RiveRef } from 'rive-react-native';
 
-const STATE_MACHINE_NAME = 'Button_Animation';
+const STATE_MACHINE_NAME = 'Button_Animation' as const;
 
 const DarklightModeAnimation = ({
   mode,
@@ -12,6 +14,16 @@ const DarklightModeAnimation = ({
   style?: ViewStyle;
 }) => {
   const riveRef = useRef<RiveRef>(null);
+
+  const onStateChanged = useCallback(
+    (state: string, event: string) => {
+      if (riveRef.current && includes(['Day/Night_Click', 'Night/Day_Click'], event)) {
+        console.log('State changed', state, event);
+        Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
+      }
+    },
+    [riveRef.current],
+  );
 
   useEffect(() => {
     if (mode === 'light') {
@@ -30,7 +42,8 @@ const DarklightModeAnimation = ({
       resourceName="dark_light_mode_switch"
       stateMachineName={STATE_MACHINE_NAME}
       style={style}
-    // url="https://public.rive.app/community/runtime-files/8771-16784-darklight-mode-switch.riv"
+      onStateChanged={onStateChanged}
+      // url="https://public.rive.app/community/runtime-files/8771-16784-darklight-mode-switch.riv"
     />
   );
 };